dailybuid: sceneに属さないsolidがHapticEngineにあることがあるので、その場合にはnullを返すようにする。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7110 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Graphics/GRDeviceGL.cpp
@@ -212,7 +212,7 @@ void GRDeviceGL::SetBlendMatrix(const Affinef& afb, unsigned int id){
 	if(id < blendMatrices.size()){
 		blendMatrices[id] = afb;
 		if(curShader)
-			glUniformMatrix4fv(curShader->blendMatricesLoc, blendMatrices.size(), false, (float*)&blendMatrices[0]);
+			glUniformMatrix4fv(curShader->blendMatricesLoc, (int)blendMatrices.size(), false, (float*)&blendMatrices[0]);
 	}
 }		
 /// 頂点フォーマットの指定
